requirements.txt
openai
tiktoken
requests
python-dotenv
streamlit
# dotenv